Tested new release 2.0: 1) Update and fix bugs 2) Added new script in order to select the shiny odds 3) Increased bag items size 4) Modified omniscient flag for all boss trainer that unlock level caps

+++ src/rom_header_gf.c
@@ -80,7 +80,7 @@ struct GFRomHeader
     u32 gcnLinkFlagsOffset;
     u32 gameClearFlag;
     u32 ribbonFlag;
-    u8 bagCountItems;
+    u16 bagCountItems;
     u8 bagCountTeraShards;
     u8 bagCountMegaStones;
     u8 bagCountKeyItems;
